requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
autogen-agentchat==0.2.40
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
Django==4.2.20
docker==7.1.0
durationpy==0.9
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.112.0
filelock==3.17.0
filetype==1.2.0
FLAML==2.3.4
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.10.0
googleapis-common-protos==1.67.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kubernetes==32.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.147
litellm==1.61.15
lxml==5.3.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
markdownify==1.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
mysqldump==0.0.10
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.63.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pi_heif==0.21.0
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.13.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.3.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.2
pytest-alembic==0.11.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.8
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.13.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.14
tokenizers==0.21.0
tomli==2.0.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.20
unstructured-client==0.30.1
unstructured-inference==0.8.9
unstructured.pytesseract==0.3.13
urllib3==2.2.2
uvicorn==0.30.5
vnstock==3.2.6
watchfiles==1.0.4
watermark==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.9.4
zipp==3.21.0
